sales-synergy-analyzer/requirements.txt
fastapi
uvicorn
yfinance>=1.0
anthropic
httpx
python-dotenv
beautifulsoup4
lxml
pydantic
